extract max decimals logic to maxDecimals.ts
